requirements.txt
absl-py==0.6.1
astor==0.7.1
gast==0.2.1
grpcio==1.17.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.4
Pillow==5.4.1
protobuf==3.6.1
PyYAML==3.13
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==0.14.1
